only show user created info for all calls

+++ routes/todo.js
@@ -7,7 +7,11 @@ const todoRoutes = Router();
 todoRoutes.get("/api/todos", verifyAuth, (req, res) => {
   const currentUser = req.user;
   return db.todo
-    .findAll()
+    .findAll({
+      where: {
+        user_id: currentUser.id,
+      },
+    })
     .then((todo) => res.send(todo))
     .catch((err) => {
       console.log("error occurred", JSON.stringify(err));
@@ -16,10 +20,16 @@ todoRoutes.get("/api/todos", verifyAuth, (req, res) => {
 });
 
 todoRoutes.post("/api/todos", verifyAuth, (req, res) => {
+  const currentUser = req.user;
+
   const { name, description } = req.body;
 
+  console.log("body", req.body);
+
+  console.log("i am current user", currentUser);
+
   return db.todo
-    .create({ name, description })
+    .create({ name, description, user_id: currentUser.id })
     .then(() => res.send(todo))
     .catch((err) => res.send(err));
 });
@@ -27,6 +37,7 @@ todoRoutes.post("/api/todos", verifyAuth, (req, res) => {
 todoRoutes.put("/api/todos/:id", verifyAuth, (req, res) => {
   const id = parseInt(req.params.id);
   const { name, description } = req.body;
+  const currentUser = req.user;
 
   const payload = {
     name,
@@ -35,7 +46,7 @@ todoRoutes.put("/api/todos/:id", verifyAuth, (req, res) => {
 
   return db.todo
     .update(payload, {
-      where: { id },
+      where: { id, user_id: currentUser.id },
     })
     .then((todo) => res.send(todo))
     .catch((err) => res.send(err));
@@ -43,9 +54,10 @@ todoRoutes.put("/api/todos/:id", verifyAuth, (req, res) => {
 
 todoRoutes.delete("/api/todos/:id", verifyAuth, (req, res) => {
   const id = parseInt(req.params.id);
+  const currentUser = req.user;
 
   return db.todo
-    .destroy({ where: { id } })
+    .destroy({ where: { id, user_id: currentUser.id } })
     .then(() => res.send({}))
     .catch((err) => res.send(err));
 });
